requirements.txt
h5py==2.10.0
Keras-Preprocessing==1.1.0
tensorflow-cpu
streamlit==0.74.1
numpy
pillow
